Sample code 부분에 대한 reorganization 제안
위와 같은 형태를 사용해보면 어떨것 같다고 생각해? 네가 생각하고 있는
샘플 부분만 따로 떼어놓고 보는데 더 좋지 않을까 생각해서 한번
조정해봤어.
